requirements.txt
pyserial
flask
flask_restful
